Move the WebComponent/LitElement that was inside the story file to own file
ESBuild errors was not happy it was inside a story file
